REFACTORING: Item component file split by subcomponents.
